cleaned up some warnings.

+++ simple_nes/src/PPU.cpp
@@ -254,9 +254,9 @@ namespace sn
                     m_cycle = 0;
                     m_pipelineState = VerticalBlank;
 
-                    for (int x = 0; x < m_pictureBuffer.size(); ++x)
+                    for (size_t x = 0; x < m_pictureBuffer.size(); ++x)
                     {
-                        for (int y = 0; y < m_pictureBuffer[0].size(); ++y)
+                        for (size_t y = 0; y < m_pictureBuffer[0].size(); ++y)
                         {
                             m_screen.setPixel(x, y, m_pictureBuffer[x][y]);
                         }
